refactor streamlit, add streamlit rotation file logging, add predict_csv implementation
