+++ inference/streamlit/pages/fit.py
@@ -1,30 +1,20 @@
 import streamlit as st
 import time
-import pandas as pd
-import numpy as np
-from sklearn.model_selection import KFold
-from sklearn.metrics import accuracy_score
-from catboost import CatBoostClassifier
-from sklearn.linear_model import RidgeClassifier
-from sklearn.preprocessing import LabelEncoder
 
-from client import ModelsAPIClient, DataAPIClient
+from client import ModelsAPIClient
 from utils import get_ridge_params, get_catboost_params
-import logging
+import streamlit_logging
 
-# Настройка логирования
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+logger = streamlit_logging.get_logger("fit_page")
 
 
-def fit_model(api_client: ModelsAPIClient):
-    """Функция для обучения модели."""
+def run_fit_page():
     st.header("Обучение модели")
     logger.info("Инициализация процесса обучения модели.")
 
-    model_type_input = st.selectbox(
-        "Выберите модель", ["⚖️ Ridge Classifier", "🧠 CatBoost Classifier"]
-    )
+    models_api_client: ModelsAPIClient = st.session_state.models_api_client
+
+    model_type_input = st.selectbox("Выберите модель", ["⚖️ Ridge Classifier", "🧠 CatBoost Classifier"])
     logger.info(f"Выбрана модель: {model_type_input}")
 
     st.subheader("Гиперпараметры модели")
@@ -50,12 +40,12 @@ def fit_model(api_client: ModelsAPIClient):
         # Запуск обучения модели
         logger.info("Запуск обучения модели.")
         start_time = time.time()
-        fit_response = api_client.fit_model(fit_request)
+        fit_response = models_api_client.fit_model(fit_request)
         st.success("Обучение модели начато!")
 
         # Проверка статуса обучения
         while True:
-            status_response = api_client.get_fit_status(model_id)
+            status_response = models_api_client.get_fit_status(model_id)
             st.write(f"Статус обучения: {status_response['status']}")
             logger.info(f"Текущий статус обучения: {status_response['status']}")
             if status_response["status"] in ["Success", "Failed"]:
@@ -64,6 +54,8 @@ def fit_model(api_client: ModelsAPIClient):
 
         end_time = time.time()
         st.write(f"⏳ Время обучения составило: {end_time - start_time:.2f} сек")
-        logger.info(
-            f"Обучение модели завершено. Время обучения: {end_time - start_time:.2f} секунд."
-        )
+        logger.info(f"Обучение модели завершено. Время обучения: {end_time - start_time:.2f} секунд.")
+
+
+if __name__ == "__page__":
+    run_fit_page()
