+++ inference/streamlit/pages/model_info.py
@@ -1,24 +1,27 @@
 import streamlit as st
-from client import ModelsAPIClient, DataAPIClient
+
+from client import ModelsAPIClient
+import streamlit_logging
+
+logger = streamlit_logging.get_logger("model_info_page")
 
 
-def display_model_info(api_client):
-    """Функция для отображения информации о модели."""
+def run_model_info_page():
     st.header("Информация о модели")
 
+    models_api_client: ModelsAPIClient = st.session_state.models_api_client
+
     if st.button("Загрузить список обученных моделей"):
-        model_list_response = api_client.get_model_list()
+        model_list_response = models_api_client.get_model_list()
         if model_list_response:
-            st.session_state.models = model_list_response[
-                "models"
-            ]  # Сохраняем список моделей
+            st.session_state.models = model_list_response["models"]  # Сохраняем список моделей
             st.write("Обученные модели:")
             st.write(st.session_state.models)
 
     model_id_input = st.selectbox("Выберите ID модели", st.session_state.models)
 
     if st.button("📖 Получить информацию о модели"):
-        model_info = api_client.get_model_info(model_id_input)
+        model_info = models_api_client.get_model_info(model_id_input)
         if model_info:
             st.write("📝 Информация о модели:")
             st.json(model_info)
@@ -26,5 +29,9 @@ def display_model_info(api_client):
             st.error("❌ Эта модель не существует.")
 
     if st.button("Активировать выбранную модель"):
-        activate_response = api_client.activate_model(model_id_input)
+        activate_response = models_api_client.activate_model(model_id_input)
         st.write("Модель активирована")
+
+
+if __name__ == "__page__":
+    run_model_info_page()
