+++ inference/streamlit/streamlit.py
@@ -1,58 +1,47 @@
 import streamlit as st
 from client import ModelsAPIClient, DataAPIClient
-from fit import fit_model
-from predict import predict_model
-from model_info import display_model_info
-from eda_streamlit import run_eda_streamlit
-import logging
+from config import get_config
+import streamlit_logging
 
-# Настройка логирования
-logging.basicConfig(level=logging.INFO)
-logger = logging.getLogger(__name__)
+logger = streamlit_logging.get_logger(__name__)
+
+
+def init_streamlit_app():
+    st.set_page_config(
+        page_title="Предсказательные модели для DotA 2",
+        page_icon="📊",
+        layout="wide",
+        initial_sidebar_state="auto",
+    )
+    if "models_api_client" not in st.session_state:
+        st.session_state.models_api_client = ModelsAPIClient(
+            get_config().client_config.fastapi_host, get_config().client_config.fastapi_port
+        )
+    if "data_api_client" not in st.session_state:
+        st.session_state.data_api_client = DataAPIClient(
+            get_config().client_config.fastapi_host, get_config().client_config.fastapi_port
+        )
+    if "models" not in st.session_state:
+        st.session_state.models = []
+    if "account_ids" not in st.session_state:
+        try:
+            account_ids_response = st.session_state.data_api_client.get_account_ids()
+            st.session_state.account_ids = account_ids_response["account_ids"]
+        except Exception as e:
+            logger.error(f"Ошибка при получении account_ids: {e}")
 
 
 def main():
-    # Инициализация API
-    host = "http://fastapi"  # Замените на рабочий хост
-    port = 8000 # Замените на рабочий порт
-    models_api_client = ModelsAPIClient(host, port)
-    data_api_client = DataAPIClient(host, port)
+    init_streamlit_app()
 
-    logger.info("Инициализация API клиентов завершена.")
-
-    # Заголовок приложения
-    st.title("Модель по анализу данных")
-
-    # Инициализация состояния сессии
-    if "page" not in st.session_state:
-        st.session_state.page = "📊 EDA"
-    if "models" not in st.session_state:
-        st.session_state.models = []
-
-    # Создание вертикального меню с кнопками
-    st.sidebar.header("Меню быстрого доступа")
-    if st.sidebar.button("📊 EDA"):  # Кнопка EDA
-        st.session_state.page = "📊 EDA"
-        logger.info("Пользователь выбрал страницу EDA.")
-    if st.sidebar.button("🔄 Обучение модели"):
-        st.session_state.page = "🔄 Обучение модели"
-        logger.info("Пользователь выбрал страницу обучения модели.")
-    if st.sidebar.button("ℹ️ Информация о модели"):
-        st.session_state.page = "ℹ️ Информация о модели"
-        logger.info("Пользователь выбрал страницу информации о модели.")
-    if st.sidebar.button("🔮 Предсказания"):
-        st.session_state.page = "🔮 Предсказания"
-        logger.info("Пользователь выбрал страницу предсказаний.")
-
-    # Определение действия на основе текущей страницы
-    if st.session_state.page == "📊 EDA":
-        run_eda_streamlit()
-    elif st.session_state.page == "🔄 Обучение модели":
-        fit_model(models_api_client)
-    elif st.session_state.page == "🔮 Предсказания":
-        predict_model(models_api_client, data_api_client)
-    elif st.session_state.page == "ℹ️ Информация о модели":
-        display_model_info(models_api_client)
+    pages = [
+        st.Page("pages/eda.py", title="EDA", icon="📊", url_path="/eda"),
+        st.Page("pages/fit.py", title=" Обучение модели", icon="🧠", url_path="/fit"),
+        st.Page("pages/model_info.py", title="Информация о модели", icon="📖", url_path="/model_info"),
+        st.Page("pages/predict.py", title="Предсказания", icon="🔮", url_path="/predict"),
+    ]
+    navigation = st.navigation(pages)
+    navigation.run()
 
 
 if __name__ == "__main__":
