+++ requirements.txt
@@ -1,10 +1,9 @@
-streamlit
-requests
-pandas
-numpy
-scikit-learn
-catboost
-fastapi[standard]
-uvicorn
-pydantic-settings
-pydantic
+streamlit==1.41.1
+pandas==2.2.3
+numpy==1.26.0
+scikit-learn==1.6.0
+catboost==1.2.7
+fastapi[standard]==0.115.5
+uvicorn==0.32.1
+pydantic-settings==2.7.1
+pydantic==2.9.2
